refactor: move finding slot tags in template text parser to separate function
